refactor: address PR review feedback for raise dispute flow
Extract RaiseDisputeDialog into its own component and clean up the disputes API route imports per maintainer review.
